* 'master' of https://github.com/AY2324S2-CS2113T-W14-3/tp:
  Add PPP `sxfoo.md`
  Create thawtunzan.md
  Edit Command Summary in UG
  Modify EXPECTED.TXT for IO test
  Add argument format in help message
  Allow JsonManager class to filter out mods in the json that do not have data on the available semester
  Add implementation that helps take in non-whole number MC mods *Example: LL4361Z
  Add skeleton for dextboy PPP
  Add skeleton for dextboy PPP
  Fix for gradlecheck
  Bug Fix with the json file
  Add minor changes to test bug
  Standardise removal of periods in UG
  Amend to adhere to gradle check
  Accept "S" as a grade now
  Add bye command and exceptions to note in UG
  Desiredgpa will catch the case if user has already has taken more than 160MCs in the planner
  Gpa will only print up to 2 decimal places
  Implement a check that ensure the module being added is correctly added to the correct semester
  Fix a bug where two different error statements were being printed when adding a module and there are duplicates

+++ src/main/java/seedu/duke/command/ViewCommand.java
@@ -13,8 +13,8 @@ public class ViewCommand extends Command {
 
     private String name = user.getName();
     private int currSem = user.getCurrentSemester();
-    private int takenMCs = moduleList.calculateTakenMCs();
-    private int totalMCs = moduleList.calculateTotalMCs();
+    private float takenMCs = moduleList.calculateTakenMCs();
+    private float totalMCs = moduleList.calculateTotalMCs();
     private Map<Integer, ArrayList<Module>> modulesBySemMap = moduleList.groupModulesBySemester();
     private final Map<String, String> args;
